requirements.txt
# --- image analysis ---
insightface==0.7.3
absl-py==2.4.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.32
nudenet==3.4.2
onnxruntime-gpu==1.24.3
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
tqdm==4.67.3

# --- LLM API (text analysis) ---
anthropic==0.84.0
google-genai==1.66.0

# --- rich (progress / console) ---
markdown-it-py==4.0.0
mdurl==0.1.2
pygments==2.19.2
rich==14.3.3

# --- selenium / browser automation ---
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
h11==0.16.0
idna==3.11
MouseInfo==0.1.3
numpy==2.4.2
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
PyAutoGUI==0.9.54
pycparser==3.0
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytweening==1.2.0
pytz==2026.1.post1
requests==2.32.5
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.3.2

ruff==0.15.5